feat(components): extract layout skeletons with injectable renderPageContent
