+++ packages/components/src/templates/next/layouts/ArticleSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { ArticleLayoutSkeleton } from "./ArticleSkeleton"
