+++ packages/components/src/templates/next/layouts/Content/Content.tsx
@@ -1,78 +1,10 @@
 import type { ContentPageSchemaType } from "~/types"
-import { tv } from "~/lib/tv"
-import { getBreadcrumbFromSiteMap } from "~/utils/getBreadcrumbFromSiteMap"
-import { getSiderailFromSiteMap } from "~/utils/getSiderailFromSiteMap"
-import { getTableOfContents } from "~/utils/getTableOfContents"
-import { getTransformedPageContent } from "~/utils/getTransformedPageContent"
 
-import { BackToTopLink } from "../../components/internal/BackToTopLink"
-import { ContentPageHeader } from "../../components/internal/ContentPageHeader"
-import { Siderail } from "../../components/internal/Siderail"
-import { TableOfContents } from "../../components/internal/TableOfContents"
 import { renderPageContent } from "../../render"
-import { Skeleton } from "../Skeleton"
+import { ContentLayoutSkeleton } from "../ContentSkeleton"
 
-const createContentLayoutStyles = tv({
-  slots: {
-    container:
-      "mx-auto grid max-w-screen-xl grid-cols-12 px-6 py-12 md:px-10 md:py-16 lg:gap-6 xl:gap-10",
-    siderailContainer: "relative col-span-3 hidden lg:block",
-    content:
-      "col-span-12 flex flex-col gap-16 break-words lg:col-span-9 lg:mr-24",
-  },
-})
-
-const compoundStyles = createContentLayoutStyles()
-
-export const ContentLayout = ({
-  site,
-  page,
-  layout,
-  content,
-}: ContentPageSchemaType) => {
-  const isParentPageRoot = page.permalink.split("/").length === 2
-
-  // Note: We do not show side rail for first-level pages
-  const sideRail = !isParentPageRoot
-    ? getSiderailFromSiteMap(site.siteMap, page.permalink)
-    : null
-
-  // auto-inject ids for heading level 2 blocks if does not exist
-  const transformedContent = getTransformedPageContent(content)
-  const tableOfContents = getTableOfContents(site, transformedContent)
-  const breadcrumb = getBreadcrumbFromSiteMap(
-    site.siteMap,
-    page.permalink.split("/").slice(1),
-  )
+export const ContentLayout = (props: ContentPageSchemaType) => {
   return (
-    <Skeleton site={site} page={page} layout={layout}>
-      <ContentPageHeader
-        {...page.contentPageHeader}
-        image={page.image}
-        title={page.title}
-        breadcrumb={breadcrumb}
-        site={site}
-        lastUpdated={page.lastModified}
-      />
-      <div className={compoundStyles.container()}>
-        <div className={compoundStyles.content()}>
-          {tableOfContents.length > 1 && (
-            <TableOfContents items={tableOfContents} />
-          )}
-          <div>
-            {renderPageContent({
-              content: transformedContent,
-              layout,
-              site,
-              permalink: page.permalink,
-            })}
-          </div>
-        </div>
-        <div className={compoundStyles.siderailContainer()}>
-          {sideRail && <Siderail {...sideRail} />}
-          <BackToTopLink />
-        </div>
-      </div>
-    </Skeleton>
+    <ContentLayoutSkeleton {...props} renderPageContent={renderPageContent} />
   )
 }
