+++ packages/components/src/templates/next/layouts/ContentSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { ContentLayoutSkeleton } from "./ContentSkeleton"
