+++ packages/components/src/templates/next/layouts/DatabaseSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { DatabaseLayoutSkeleton } from "./DatabaseSkeleton"
