+++ packages/components/src/templates/next/layouts/Homepage/Homepage.tsx
@@ -1,28 +1,10 @@
 import { type HomePageSchemaType } from "~/types"
 
 import { renderPageContent } from "../../render"
-import { Skeleton } from "../Skeleton"
+import { HomepageLayoutSkeleton } from "../HomepageSkeleton"
 
-export const HomepageLayout = ({
-  site,
-  page,
-  layout,
-  content,
-}: HomePageSchemaType) => {
+export const HomepageLayout = (props: HomePageSchemaType) => {
   return (
-    <Skeleton site={site} page={page} layout={layout}>
-      <div
-        // ComponentContent = "component-content" (customCssClass.ts) is imported by all Homepage components,
-        // but cannot be used here as tailwind does not support dynamic class names
-        className={`break-words [&_.component-content]:mx-auto [&_.component-content]:max-w-screen-xl [&_.component-content]:px-6 [&_.component-content]:md:px-10`}
-      >
-        {renderPageContent({
-          content,
-          layout,
-          site,
-          permalink: page.permalink,
-        })}
-      </div>
-    </Skeleton>
+    <HomepageLayoutSkeleton {...props} renderPageContent={renderPageContent} />
   )
 }
