+++ packages/components/src/templates/next/layouts/HomepageSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { HomepageLayoutSkeleton } from "./HomepageSkeleton"
