+++ packages/components/src/templates/next/layouts/IndexPageSkeleton/IndexPageSkeleton.tsx
@@ -1,0 +1,75 @@
+import type { IndexPageSchemaType } from "~/types"
+import { tv } from "~/lib/tv"
+import { getBreadcrumbFromSiteMap } from "~/utils/getBreadcrumbFromSiteMap"
+import { getTableOfContents } from "~/utils/getTableOfContents"
+import { getTransformedPageContent } from "~/utils/getTransformedPageContent"
+
+import type {
+  RenderPageContentOutput,
+  RenderPageContentParams,
+} from "../../render/types"
+import { ContentPageHeader } from "../../components/internal/ContentPageHeader"
+import { TableOfContents } from "../../components/internal/TableOfContents"
+import { ensureChildrenPagesBlock } from "../IndexPage/ensureChildrenPagesBlock"
+import { Skeleton } from "../Skeleton"
+
+const createIndexPageLayoutStyles = tv({
+  slots: {
+    container:
+      "mx-auto grid max-w-screen-xl grid-cols-12 px-6 py-12 md:px-10 md:py-16 lg:gap-10",
+    siderailContainer: "relative col-span-3 hidden lg:block",
+    content: "col-span-12 break-words lg:col-span-8",
+  },
+})
+
+const compoundStyles = createIndexPageLayoutStyles()
+
+interface IndexPageLayoutSkeletonProps extends IndexPageSchemaType {
+  renderPageContent: (
+    params: RenderPageContentParams,
+  ) => RenderPageContentOutput
+}
+
+export const IndexPageLayoutSkeleton = ({
+  site,
+  page,
+  layout,
+  content,
+  renderPageContent,
+}: IndexPageLayoutSkeletonProps) => {
+  const breadcrumb = getBreadcrumbFromSiteMap(
+    site.siteMap,
+    page.permalink.split("/").slice(1),
+  )
+
+  const pageContent = ensureChildrenPagesBlock(content)
+  // auto-inject ids for heading level 2 blocks if does not exist
+  const transformedContent = getTransformedPageContent(pageContent)
+  const tableOfContents = getTableOfContents(site, transformedContent)
+
+  return (
+    <Skeleton site={site} page={page} layout={layout}>
+      <ContentPageHeader
+        {...page.contentPageHeader}
+        colorScheme="inverse"
+        title={page.title}
+        breadcrumb={breadcrumb}
+        site={site}
+        lastUpdated={page.lastModified}
+      />
+      <div className={compoundStyles.container()}>
+        <div className={compoundStyles.content()}>
+          {tableOfContents.length > 1 && (
+            <TableOfContents items={tableOfContents} />
+          )}
+          {renderPageContent({
+            content: transformedContent,
+            layout,
+            site,
+            permalink: page.permalink,
+          })}
+        </div>
+      </div>
+    </Skeleton>
+  )
+}
