+++ packages/components/src/templates/next/layouts/IndexPageSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { IndexPageLayoutSkeleton } from "./IndexPageSkeleton"
