+++ packages/components/src/templates/next/render/index.ts
@@ -1,4 +1,11 @@
 export { renderLayout } from "./renderLayout"
 export { renderPageContent } from "./renderPageContent"
+export { renderPageContentSkeleton } from "./renderPageContentSkeleton"
 export { renderComponentPreviewText } from "./renderComponentPreviewText"
 export { renderPrefillText } from "./renderPrefillText"
+export type {
+  RenderComponentOutput,
+  RenderComponentProps,
+  RenderPageContentOutput,
+  RenderPageContentParams,
+} from "./types"
