+++ packages/components/src/templates/next/render/renderPageContent.ts
@@ -1,60 +1,7 @@
-import type {
-  IsomerComponent,
-  IsomerPageLayoutType,
-  IsomerSiteProps,
-} from "~/types"
-
-import { doesComponentHaveImage } from "./doesComponentHaveImage"
+import type { RenderPageContentParams } from "./types"
 import { renderComponent } from "./renderComponent"
+import { renderPageContentSkeleton } from "./renderPageContentSkeleton"
 
-interface RenderPageContentParams {
-  content: IsomerComponent[]
-  layout: IsomerPageLayoutType
-  site: IsomerSiteProps
-  permalink: string
-}
-
-export const renderPageContent = ({
-  content,
-  ...rest
-}: RenderPageContentParams) => {
-  // Filter out hidden childrenpages blocks
-  const visibleContent = content.filter((component) =>
-    component.type === "childrenpages" ? !component.isHidden : true,
-  )
-
-  // Find index of first component with image
-  const firstImageIndex = visibleContent.findIndex((component) =>
-    doesComponentHaveImage({ component }),
-  )
-
-  let isInfopicTextOnRight = false
-
-  return visibleContent.map((component, index) => {
-    // Lazy load components with images that appear after the first image.
-    // We assume that only the first image component will be visible above the fold,
-    // while subsequent components should be lazy loaded to enhance the Lighthouse performance score.
-    const shouldLazyLoad = index > firstImageIndex
-
-    if (component.type === "infopic") {
-      isInfopicTextOnRight = !isInfopicTextOnRight
-      const formattedComponent = {
-        ...component,
-        isTextOnRight: isInfopicTextOnRight,
-      }
-      return renderComponent({
-        elementKey: index,
-        component: formattedComponent,
-        shouldLazyLoad,
-        ...rest,
-      })
-    }
-
-    return renderComponent({
-      elementKey: index,
-      component,
-      shouldLazyLoad,
-      ...rest,
-    })
-  })
+export const renderPageContent = (params: RenderPageContentParams) => {
+  return renderPageContentSkeleton({ ...params, renderComponent })
 }
