+++ packages/components/src/templates/next/render/renderPageContentSkeleton.ts
@@ -1,0 +1,57 @@
+import type {
+  RenderComponentOutput,
+  RenderComponentProps,
+  RenderPageContentOutput,
+  RenderPageContentParams,
+} from "./types"
+import { doesComponentHaveImage } from "./doesComponentHaveImage"
+
+interface RenderPageContentSkeletonProps extends RenderPageContentParams {
+  renderComponent: (props: RenderComponentProps) => RenderComponentOutput
+}
+
+export const renderPageContentSkeleton = ({
+  renderComponent,
+  content,
+  ...rest
+}: RenderPageContentSkeletonProps): RenderPageContentOutput => {
+  // Filter out hidden childrenpages blocks
+  const visibleContent = content.filter((component) =>
+    component.type === "childrenpages" ? !component.isHidden : true,
+  )
+
+  // Find index of first component with image
+  const firstImageIndex = visibleContent.findIndex((component) =>
+    doesComponentHaveImage({ component }),
+  )
+
+  let isInfopicTextOnRight = false
+
+  return visibleContent.map((component, index) => {
+    // Lazy load components with images that appear after the first image.
+    // We assume that only the first image component will be visible above the fold,
+    // while subsequent components should be lazy loaded to enhance the Lighthouse performance score.
+    const shouldLazyLoad = index > firstImageIndex
+
+    if (component.type === "infopic") {
+      isInfopicTextOnRight = !isInfopicTextOnRight
+      const formattedComponent = {
+        ...component,
+        isTextOnRight: isInfopicTextOnRight,
+      }
+      return renderComponent({
+        elementKey: index,
+        component: formattedComponent,
+        shouldLazyLoad,
+        ...rest,
+      })
+    }
+
+    return renderComponent({
+      elementKey: index,
+      component,
+      shouldLazyLoad,
+      ...rest,
+    })
+  })
+}
